adds extension method for .ToEpoch()

+++ src/Portfolio.Tests/Lib/DateTimeExtensionsTests.cs
@@ -8,7 +8,7 @@ namespace Portfolio.Lib
     {
         [Test]
         [TestCase(2013, 12, 13, 10, 32, 45, Result = 1386930765)]
-        public int ToEpoch_returns_expected_result(int year, int month, int day, int hour, int minute, int second)
+        public long ToEpoch_returns_expected_result(int year, int month, int day, int hour, int minute, int second)
         {
             var dateTime = new DateTime(year, month, day, hour, minute, second);
             var epoch = dateTime.ToEpoch();
